requirements_preview.txt
streamlit
faiss-cpu
pandas
numpy
torch
transformers
datasets
